% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acnrlab_instruments.R
\docType{data}
\name{acnrlab_instruments}
\alias{acnrlab_instruments}
\title{Instrument Lookup Table}
\format{
A data frame with 18 rows and 4 variables:
\describe{
\item{instrument_name}{Name of the instrument}
\item{instrument_prefix}{Instrument prefix in variable names}
\item{fields}{list of variables/field names for this instrument}
\item{is_repeating}{True if it's a repeating instrument, else False}
}
}
\usage{
acnrlab_instruments
}
\description{
Instrument information for our database for use with various functions
}
\keyword{datasets}
