% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_from_demographics.R
\name{add_from_demographics}
\alias{add_from_demographics}
\title{Add Demographics Information to Info Table}
\usage{
add_from_demographics(
  extracted_info,
  demographics,
  include_cols,
  record_id_col = "record_id"
)
}
\arguments{
\item{extracted_info}{Output of extract_info}

\item{demographics}{Demographics table from split_instruments}

\item{include_cols}{Which columns to add to the info table}

\item{record_id_col}{Defaults to record_id}
}
\value{
The extracted_info dataframe with the columns from demographics
specified in include_cols
}
\description{
This is a wrapper for left_join to make it a bit easier for people without
knowledge of SQL-esque joins to simply take information from the demographics
table (such as subject ID or aphasia status) and add it to the info table.
}
