% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_subjects.R
\name{.filter_non_repeating}
\alias{.filter_non_repeating}
\title{Filter non repeating instruments}
\usage{
.filter_non_repeating(instrument_db, filter_specification)
}
\arguments{
\item{instrument_db}{Unfiltered list of instruments}

\item{filter_specification}{Dataframe, the output of extract_info}
}
\value{
List of non-repeating instruments with only the records specified in
the filter specification
}
\description{
Demographic info, and other instruments which aren't repeating instruments,
aren't filtered in the same procedure as the other instruments since there's
no specification of timepoint or study in them. To remove any records we
don't need, we can filter by which record_ids remain in the filter specification
passed to query_subjects.
}
