% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_instrument.R
\name{.get_nonrepeating_insts}
\alias{.get_nonrepeating_insts}
\title{Process non repeating instruments}
\usage{
.get_nonrepeating_insts(inst, lookup, redcap_cols, verbose)
}
\arguments{
\item{inst}{Dataframe of a single instrument, this will typically be the
output of filtering the report where redcap_repeating_instrument is NA}

\item{lookup}{Lookup table, passed from split_instruments}

\item{redcap_cols}{Redcap metadata column names, passed from split_instruments}

\item{verbose}{Logical, whether readr type conversion messsages will be printed}
}
\value{
List of tibbles for each non repeating instrument found in inst
}
\description{
Used to handle the demographic and medical info, mostly
}
