% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{.has_info_in_record}
\alias{.has_info_in_record}
\title{Check if record contains any info variables}
\usage{
.has_info_in_record(record_df, pivot_cols)
}
\arguments{
\item{record_df}{Pivoted dataframe with extracted info variables}

\item{pivot_cols}{Which info columns the df was pivoted by}
}
\value{
Logical vector of which rows to keep
}
\description{
Given a dataframe that was pivoted using info variables stored in pivot_cols,
return only the rows that have at least 1 info variable filled out. Essentially,
check each row and return TRUE if any info column is not NA. This is necessary
because not all info variables will be included in every report.
}
