% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_instrument.R
\name{.has_instrument}
\alias{.has_instrument}
\title{Check if instrument in report df}
\usage{
.has_instrument(inst, check_prefixes, USE.NAMES = FALSE)
}
\arguments{
\item{inst}{Dataframe of a single instrument, this will typically be the
output of filtering the report where redcap_repeating_instrument is NA}

\item{check_prefixes}{Which prefixes to check, typically c('demo','med')}

\item{USE.NAMES}{Whether to return prefixes as names in the logical vector,
defaults to FALSE}
}
\value{
Logical vector, TRUE if prefix is found in the colnames of the instrument
df, FALSE if not
}
\description{
Given a character vector of instrument prefixes (typically non-repeating
instruments like demo and med), return whether a variable from those instruments
is included. This is equivalent to something like
grepl(colnames(inst), "^demo_") but will return as soon as it finds one. So,
best case O(3) worst case O(n)
}
