% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_instrument.R
\name{.pivot_instrument}
\alias{.pivot_instrument}
\title{Pivot instrument}
\usage{
.pivot_instrument(
  instrument,
  instrument_prefix,
  instrument_regex = "_(.+)_(.+)_(.+)$",
  output_names = c("subtest", "number", ".value"),
  ...
)
}
\arguments{
\item{instrument}{One instrument's data}

\item{instrument_prefix}{Instrument prefix}

\item{instrument_regex}{The regex associated with the metadata in the field
name}

\item{output_names}{Output names for the capture groups of the regex}

\item{...}{Other options for pivot_longer()}
}
\value{
A pivoted instrument
}
\description{
Helper for trim_instrument, handles the actual pivot across particular
columns. NOTE: The instrument_regex parameter is what our lab uses across
most instruments. If this needs to be changed later or generalizable for
other labs, then it's there. For now things work for us just fine.
}
\examples{
# TODO
}
