% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{.preprocess}
\alias{.preprocess}
\title{Preprocess data}
\usage{
.preprocess(inst_data, inst_prefix, inst_name)
}
\arguments{
\item{inst_data}{A single instrument's data}

\item{inst_prefix}{The prefix for this instrument}

\item{inst_name}{The name for this instrument}
}
\value{
A smaller dataframe with some columns renamed depending on the
instrument provided
}
\description{
Our field names in Redcap have a lot of information in them, but some of them
don't need to be used by the researcher at all. This removes some redundant
information that's common to all instruments and also calls additional
instrument-specific functions as needed. For example, checkbox matrices
in redcap append ___X to the field name, but what X means will vary from
instrument to instrument.
}
\examples{
# TODO
}
