% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_subjects.R
\name{.verify_datatypes}
\alias{.verify_datatypes}
\title{Verify datatypes before joining info}
\usage{
.verify_datatypes(filter_spec, inst_data, cols)
}
\arguments{
\item{filter_spec}{Filter specification}

\item{inst_data}{Data of one instrument}

\item{cols}{Columns to verify}
}
\value{

}
\description{
Helper for query_subjects(), the column datatypes for the dataset from
an api call and importing from a file can differ, which prevents the inner
join from working.
}
