% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_sqlite.R
\name{export_sqlite}
\alias{export_sqlite}
\title{Export to SQLite Database}
\usage{
export_sqlite(conn, instrument_list)
}
\arguments{
\item{conn}{A database connection from RSQLite::dbConnect()}

\item{instrument_list}{A list of instruments, either from split_instruments()
or query_subjects() (but you probably want the full data from the former)}
}
\value{
NULL if successful, a named vector if one or more are unsuccessful
}
\description{
Add a list of instrument dataframes from split_instruments() to an RSQLite
database, if you want to use SQL commands that way. This is really just a
wrapper for lapply to add all the instruments at once.
}
\examples{

all_instruments <- split_instruments(report, lookup)
db_con <- RSQLite::dbConnect(RSQLite::SQLite(), ":memory:")

export_sqlite(db_con, all_instruments)

}
