% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_to_excel.R
\name{export_to_excel}
\alias{export_to_excel}
\title{Export instruments to excel workbook}
\usage{
export_to_excel(instrument_db, filepath)
}
\arguments{
\item{instrument_db}{List of instruments}

\item{filepath}{Filepath for a new .xlsx file. Must be .xlsx.}
}
\value{

}
\description{
Export a list of instruments (the output of \code{split_instruments} or
\code{query_instruments}) to an excel workbook. Each instrument is saved in a
separate sheet.
}
\details{
Be VERY CAREFUL with dates in Excel.

This function requires the xlsx package to be installed.
}
