% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_info}
\alias{extract_info}
\title{Extract Info Variables from Instrument}
\usage{
extract_info(
  report_data,
  lookup = acnrlab_instruments,
  make_uid = T,
  record_id_col = "record_id",
  filter_by = "user"
)
}
\arguments{
\item{report_data}{Dataframe, the raw report data from get_report or importing from .csv}

\item{lookup}{Dataframe, instrument Name-Prefix lookup table from generate_table or read in from a .csv}

\item{make_uid}{Logical, should a unique id from the record id, instrument, and instance number be made?}

\item{record_id_col}{String, name of the record id column, defaults to record_id}

\item{filter_by}{String, name of the column to filter out NAs, defaults to user}
}
\value{
A dataframe with all of the "info" data from each instrument
}
\description{
Extract the information from the top of each instrument such as age & tsonset.
This is valuable to filter which records you actually want, then you
can filter each instrument via joining operations
}
\examples{
TODO
}
