% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{generate_lookup}
\alias{generate_lookup}
\title{Generate Instrument Name-Prefix Lookup Table}
\usage{
generate_lookup(redcap_uri, redcap_token)
}
\arguments{
\item{redcap_uri}{Redcap user API URI}

\item{redcap_token}{Redcap user API token}
}
\value{

}
\description{
Must have API privileges and be connected to the Northwestern VPN. Creates
a lookup table based on the RedcapR::redcap_metadata_read() function. Note:
I need to include an offline object within the package for people to use.
}
