% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{generate_new_lookup}
\alias{generate_new_lookup}
\title{Generate Instrument Name-Prefix Lookup Table}
\usage{
generate_new_lookup(redcap_uri, redcap_token)
}
\arguments{
\item{redcap_uri}{Redcap user API URI}

\item{redcap_token}{Redcap user API token}
}
\value{
A dataframe with all the instrument names and their prefixes
}
\description{
Must have API privileges and be connected to the Northwestern VPN. This
will return a new version of the anrlab_instruments objects. Future
maintainers should run this and update that object whenever changes are made
to instruments or new instruments are added. This shouldn't happen frequently
once a project is moved to production mode.
}
