% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_redcap_utils.R
\name{get_repeating}
\alias{get_repeating}
\title{Get Repeating Instruments}
\usage{
get_repeating(redcap_uri, redcap_token)
}
\arguments{
\item{redcap_uri}{Redcap user API URI}

\item{redcap_token}{Redcap user API token}
}
\value{
A vector with repeating instrument names
}
\description{
Returns a vector with all the repeating instrument names.
Used when generating a new lookup table.
}
