% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_redcap_utils.R
\name{get_report}
\alias{get_report}
\title{Get Redcap Report}
\usage{
get_report(rep_id, redcap_uri, redcap_token)
}
\arguments{
\item{rep_id}{Report ID on Redcap}

\item{redcap_uri}{Redcap user API URI}

\item{redcap_token}{Redcap user API token}
}
\value{
A tibble representation of the specified redcap report
}
\description{
Get Redcap Report
}
