% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_redcap_csv.R
\name{load_redcap_csv}
\alias{load_redcap_csv}
\title{Load redcap csv data}
\usage{
load_redcap_csv(filepath)
}
\arguments{
\item{filepath}{csv filepath}
}
\description{
Helper function to do some prepreocessing on the data to avoid mismatching
error types due to user error.
}
