% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify_instrument.R
\name{prettify_instrument}
\alias{prettify_instrument}
\title{Prettify instrument}
\usage{
prettify_instrument(inst_data, inst_prefix)
}
\arguments{
\item{inst_data}{Instrument data}

\item{inst_prefix}{Instrument prefix}
}
\value{

}
\description{
Renames columns and applies labels as needed to make the data easier to read.
This is not called automatically like .preprocess() and .postprocess(), and
so is done on an as-needed basis.
}
