% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_subjects.R
\name{query_subjects}
\alias{query_subjects}
\title{Query Subjects Across Instruments by External Dataframe}
\usage{
query_subjects(
  instrument_db,
  filter_specification,
  join_info = T,
  limit_cols = F,
  include_cols = NA,
  record_id_col = "demo_record_id"
)
}
\arguments{
\item{instrument_db}{List of dataframes, the output from split_instruments}

\item{filter_specification}{Dataframe, the output from extract_infos which
can be filtered as desired}

\item{join_info}{Logical, defaults to True, should this function add the info
from the filter specification dataframe such as age and time since onset?}

\item{limit_cols}{Logical, assuming join_info is True, if this is False, then
include all additional columns from the filter specification. If this is True,
then include only the columns given by include_cols}

\item{include_cols}{Character Vector, vector of which columns to include.
Must be present in filter_specification.}

\item{record_id_col}{}
}
\value{

}
\description{
This function utilizes the output of extract_infos, which can then be
filtered according to some criteria, to filter the list of dataframes from
split_instruments. This lets us avoid having to utilize apply or map
functions to apply a particular filter like "Only the Project A patients
at baseline."
}
