% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_instruments.R
\name{split_instruments}
\alias{split_instruments}
\title{Split Instruments from Raw Report}
\usage{
split_instruments(
  report_data,
  lookup = acnrlab_instruments,
  record_id_col = "record_id",
  verbose = T
)
}
\arguments{
\item{report_data}{Dataframe, the raw exported data from a report}

\item{lookup}{Dataframe, instrument name-prefix lookup table.}

\item{record_id_col}{Defaults to record_id}

\item{verbose}{Logical, if True will print each instrument as it's processing}
}
\value{
List of dataframes, named according to instrument prefix
}
\description{
Given a dataframe from a raw report, convert the dataframe into a list of
dataframes. Each dataframe is 1 Redcap instrument's worth of data in tidy
format. Instrument data can be accessed through $ notation such as
output$nnb to get the data from the Northwestern Naming Battery. The name
of each dataframe is set to be the instrument prefix to keep things short.
}
\examples{
TODO
}
