% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_instruments2.R
\name{split_instruments2}
\alias{split_instruments2}
\title{Title}
\usage{
split_instruments2(
  report_data,
  lookup = inst_lookup,
  record_id_col = "record_id",
  verbose = T
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
This is the new way I'm testing out to do it...
}
