% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test dataset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 90 rows and 3792 columns.
}
\usage{
data(test_data)
}
\description{
Test dataset
}
\keyword{test_data}
