% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{.postprocess}
\alias{.postprocess}
\title{Postprocess instrument data}
\usage{
.postprocess(inst_data, inst_prefix, verbose = T)
}
\arguments{
\item{inst_data}{One instrument's data}

\item{inst_prefix}{Character, instrument's prefix}

\item{verbose}{Logical, defaults to True, inherits from split_instruments}
}
\value{
A dataframe with some cleaned up names and datatypes
}
\description{
After pivoting an instrument, some postprocessing needs to be done. Usually
this means converting datatypes from character to numeric.
}
