% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_sqlite.R
\name{export_sqlite}
\alias{export_sqlite}
\title{Export to SQLite Database}
\usage{
export_sqlite(conn, instrument_list, ...)
}
\arguments{
\item{conn}{A database connection from RSQLite::dbConnect()}

\item{instrument_list}{A list of instruments, either from split_instruments()
or query_subjects() (but you probably want the full data from the former)}
}
\value{
Nothing, will display a message if successful
}
\description{
Add a list of instrument dataframes from split_instruments() to an RSQLite
database, if you want to use SQL commands that way. This is really just a
wrapper for lapply to add all the instruments at once.
}
\examples{

# Set up instruments and database
all_instruments <- split_instruments(report, lookup)
db_con <- RSQLite::dbConnect(RSQLite::SQLite(), ":memory:")

# Send all instruments to database connection
export_sqlite(db_con, all_instruments)

# Run a simple query
RSQLite::dbGetQuery(dbcon,"SELECT
                             wab.*,
                             wabsum.aq
                           FROM wab
                           LEFT JOIN wabsum
                           USING (record_id, redcap_repeat_instance)")

}
